+++ src/main/resources/db.properties
@@ -1,6 +1,6 @@
 jdbc.driver=com.mysql.cj.jdbc.Driver
-jdbc.url=jdbc:mysql://192.168.99.100:3306/lumen?useSSL=false
+jdbc.url=jdbc:mysql://localhost:3306/lumen?useSSL=false
 jdbc.username=root
-jdbc.password=root
+jdbc.password=123456
 jdbc.max=10
 jdbc.min=1
